% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogarch_core.R
\name{ecogarch_qmle}
\alias{ecogarch_qmle}
\title{Core QML estimator for ECOGARCH(1,1) on irregularly spaced data}
\usage{
ecogarch_qmle(
  Gdt,
  dt,
  init = list(a1 = 0.1, theta = -0.1, gamma = 0.3, mu = -3),
  method = c("approx", "exact"),
  grad = c("none", "fd"),
  h = 1e-04
)
}
\arguments{
\item{Gdt}{Returns at jump times}

\item{dt}{Interarrival times}

\item{init}{list(a1, theta, gamma, mu)}

\item{method}{recursion type ('approx'|'exact')}

\item{grad}{gradient option ('none'|'fd')}

\item{h}{FD step size}
}
\description{
Core QML estimator for ECOGARCH(1,1) on irregularly spaced data
}
