% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qml.R
\name{ecogarch_qmle_mu}
\alias{ecogarch_qmle_mu}
\title{QML wrappers (mu and HF variants)}
\usage{
ecogarch_qmle_mu(
  G,
  tQML,
  init = list(a1 = 0.1, theta = -0.1, gamma = 0.3, mu = -3)
)
}
\description{
QML wrappers (mu and HF variants)
}
