% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_sigma_ecogarch}
\alias{estimate_sigma_ecogarch}
\title{Estimate volatility and innovations given ECOGARCH(1,1) parameters}
\usage{
estimate_sigma_ecogarch(par, G, tQML, method = c("approx", "exact"))
}
\arguments{
\item{par}{list(a1, theta, gamma, mu, lambda, K)}

\item{G}{log-prices}

\item{tQML}{observation times}

\item{method}{'approx'|'exact'}
}
\description{
Estimate volatility and innovations given ECOGARCH(1,1) parameters
}
