% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_matrix.R
\name{lb_residuals_matrix}
\alias{lb_residuals_matrix}
\title{Ljung-Box test on squared residuals for each column of a return matrix}
\usage{
lb_residuals_matrix(G, sigma, alpha = 0.05, lag = NULL)
}
\description{
Ljung-Box test on squared residuals for each column of a return matrix
}
