% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_one_step}
\alias{predict_one_step}
\title{One-step recursive volatility + price PI}
\usage{
predict_one_step(G, tQML, par, level = 0.95)
}
\description{
One-step recursive volatility + price PI
}
