% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_stepwise.R
\name{prediction_interval_coverage}
\alias{prediction_interval_coverage}
\title{Coverage helper}
\usage{
prediction_interval_coverage(G, G_upper, G_lower, start_idx)
}
\description{
Coverage helper
}
