% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_stepwise.R
\name{prediction_stepwise}
\alias{prediction_stepwise}
\title{Stepwise one-step prediction and coverage}
\usage{
prediction_stepwise(
  tQML,
  G,
  alpha = 0.05,
  start_k = 100,
  method = "approx",
  grad = "none"
)
}
\description{
Stepwise one-step prediction and coverage
}
