% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_cp_univariate.R
\name{simulate_ecogarch11_cp_univariate}
\alias{simulate_ecogarch11_cp_univariate}
\title{Simulate univariate ECOGARCH(1,1) driven by compound Poisson}
\usage{
simulate_ecogarch11_cp_univariate(
  par,
  lambda = 1,
  nmax = 3000,
  nu = 6,
  equidistant_step = NULL
)
}
\description{
Simulate univariate ECOGARCH(1,1) driven by compound Poisson
}
