% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midquote.R
\name{tba_midquotes}
\alias{tba_midquotes}
\title{Consolidate bid/ask quotes at identical timestamps to mid-quotes}
\usage{
tba_midquotes(tba, b, a, tgrid = NULL)
}
\description{
Consolidate bid/ask quotes at identical timestamps to mid-quotes
}
