% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{vol_prediction_density_piecewise}
\alias{vol_prediction_density_piecewise}
\title{Piecewise lognormal one-step volatility density}
\usage{
vol_prediction_density_piecewise(Xn, dt_next, par, grid = NULL, exact = TRUE)
}
\description{
Piecewise lognormal one-step volatility density
}
